Move playbooks.py and enums.py from core to data package
Restructure module organization by moving data-related modules
from osism.core to osism.data package. Updated import statements
in apply.py and validate.py to reflect the new module locations.

AI-assisted: Claude Code
