--- osism/core/__init__.py
@@ -1,1 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
